Dir: return paths in generic format
Return paths in generic pathname format (using / as the directory
separator), as opposed to the native format. The generic format is
accepted by both POSIX and Windows systems.

Using the generic format makes it simpler to construct URIs for local
files on systems where / is not the native directory separator.

Change-Id: I3baf27ebea0136066983aee6ac5f35bc3e4f0ba9
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/110343
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>