Remove CoreApplication::readFile as we already have KDUtils::File
The reason for the existence of CoreApplication::readFile was to support
Android specificities. Yet it lead to a source of breakage in libraries using
KDUtils
- A CoreApplication had to be initialized for file loading to work
- CoreApplication::readFile would throw on failure to load

To unify file handling we now only have KDUtils::File.
AndroidPlatformIntegration was update to register the AssetManager against
File so that Android support is maintained.

Change-Id: I3fc5098eb2a1e0be686632f696fa29d77b3e8b51
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/125996
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>