linux_wayland_platform_integration.h: forward declare LinuxWaylandPlatformInput
So that we don't indirectly include wayland-client-core.h and xcbcommon.h when
something includes linux_wayland_platform_integration.h.

That is because KDGui links privately to wayland and xcb. Therefore, if a library
links against KDGui and were to include wayland or xcb headers indirectly, it
would potentially fail to build since the include paths aren't added for private
libs.

Change-Id: I9f39c421b752bc6488a49593e3184d6aeb244d83
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/128176
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>