LinuxPlatformEventLoop: handle unregisterFileDescriptor on bad file as success
In case we attempted to unregister a previously registered file descriptor
while the respective file is not accessible any more (resulting in errno == EBADF),
proceed as if unregistering the file descriptor was successful.
This way we ensure notifiers are reset / deleted properly.

IMO this is a common use case when files/sockets/etc. get closed without our notice.

We previously would handle this case as an error and would keep notifiers assigned to
the file descriptor. The next time registerFileDesciptor was called on the same fd,
we would not get notified on file descriptor events anymore.

Change-Id: Ifd3901d1eb7792b41149da09585668f4a0df5af6
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/130379
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>