Add override for preferred platform integration via KDGUI_PLATFORM
It's sometimes useful to be able to switch between the two currently
available linux platform integrations: XCB and Wayland. You can set the
environment variable KDGUI_PLATFORM to "xcb" to prefer the XCB platform,
which only makes sense if you're running on Wayland because of the
presence of XWayland.

An example use case is testing the XCB backend on Wayland without having
to recompile KDUtils to stub it out.

Change-Id: I84405d3ee06e41ac231971c62d52a1af414d8691
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/132986
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Miłosz Kosobucki <milosz.kosobucki@kdab.com>