Introduce KDUtils::Logger and remove old KDUtils::Logger wrapper
This allows consumers of this library to set a factory function to create
spdlog loggers. This can be used so that the loggers are created and
registered in the spdlog registry of the application rather than the library.
This in turn allows the application to more easily have control over its
logging configuration.

See https://github.com/gabime/spdlog/wiki/How-to-use-spdlog-in-DLLs.

Once the downstream libs are ported over we can remove the
KDFoundation::createLogger() function as this will make it redundant.

Change-Id: Iedeeded1567f6cb45dbda68e6540ba71e2233d99
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/134405
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>