Improve logger helper
So that we need a factory that checks the spdlog registry in its
own dll before creating a new logger. Or, if no factory is set we
check the KDUtils registry.

Call sites should not worry about checkign via spdlog::get()
themselves as this helper and any installed factory should now
do it instead.

This will ensure the same registry is checked as is used when
creating each logger.

Change-Id: I24399e7ca778ddd310d6074300e70143ca3d2d9f
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/134715
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>