Add method to remove events targeting a specific EventReceiver
We need a mechanism to remove all events targeting a specific
EventReciever from the event queue to be able to implement
deferred deletion of Objects.
This commit adds such a mechanism by changing the underlying
event queue type from std::queue to std:deque and providing
a new method that removes events from the event queue using
the erase-remove idiom.

Change-Id: I51f90a948732990f77934731ca853b7c68018d5b
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/135489
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Miłosz Kosobucki <milosz.kosobucki@kdab.com>