Add method to defer delete Objects
This change adds new type DeferredDeleteEvent and method
Object::deleteLater(). Calling deleteLater on an object posts a
DeferredDeleteEvent to the event queue of the core application.
The DeferredDeleteEvent's EventReceiver is set to the Object to
be deleted. When the control returns to the event loop
CoreApplication::processEvents is called, the DeferredDeleteEvent
is delivered to the Object and the Object is deleted in it's
event handler.
On destruction of an EventReceiver instance all pending events
targeting the object in destruction are removed from the core
application's event queue in the EventReceiver DTOR.

In case Object::deleteLater() is called without a CoreApplication
instance, we print an error and ignore the request.
In case Object::deleteLater() is called on the CoreApplication itself,
we print an error and ignore the request since defer deleting a
CoreApplication does not make much sense.
In case the CoreApplication is deleted before the control returns
to the event loop, all pending events will now be processed in
CoreApplication's DTOR.

Change-Id: Iad6097aa23763077d88fa69f75c834bbf97a5447
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/135490
Reviewed-by: Miłosz Kosobucki <milosz.kosobucki@kdab.com>
Tested-by: Continuous Integration <build@kdab.com>