Improve mouse events API
Now mouse press/release events contain information about both the
button that caused the event (in ::button()) and about all the currently
pressed buttons (in ::buttons()).

This required implementing grabbing because otherwise, when the cursor
left the window and button was released, we never got the info about
the button release. In practice, this grabbing is needed only on
Windows.

Now we grab when the first button is pressed and release when the
last button is released.

WARNING: the semantics of ::buttons() method was changed, now it reports
all the buttons that are currently pressed consistently on all platforms
whereas before it worked only on MacOS (which tracked the button state)
albeit incorrectly.

Drive-By: Fix implementation of [KDGuiView otherMouseUp] in
cocoa_platform_window.mm which erroneously sent MousePressEvent
instead of MouseReleaseEvent.

Change-Id: I6064a851cf91097f2006e1ccf0ddecdb97d83d37
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/135916
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>