Use external fmt for spdlog to avoid MSVC warnings
The old version of fmt bundled with spdlog 1.x throws warnings on newer
MSVC.

I also bumped spdlog dependency to the latest version. We need 1.12.0
because in this version the support for fmt >=10.x was introduced.

Commit hashes are used instead of version tags to guard against
potential tampering of the dependency repository.

Change-Id: If8ac7d12efe190693524d5af38b5cdb6a4bfac8a
Reviewed-on: https://codereview.kdab.com/c/kdab/kdutils/+/136398
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Mike Krus <mike.krus@kdab.com>