Implement auto invocation of deferred slots
Connection invocations are now evaluated automatically in the event
loop. Any time an invocation is enqueued using the event loop's
connection evaluator, the event loop is woken up.

This required a KDUtils-specific ConnectionEvaluator.

When given event loop is destroyed, it sets a reference in the evaluator
to null. It's a defensive measure for the future to prevent misuse of
lingering evaluators in case the user deletes the event loop first.

AbstractEventLoop::waitForEvents is not pure abstract anymore as it
ensures evaluation of slot invocations.

Removed one copypasted comment in android event loop impl.