Fix race condition with deferred connections
Lazy creation of ConnectionEvaluator in event loop was not guarded and
it may happen from a secondary thread (or multiple of those as was the
case with tests). This is actually quite probable as the idea behind
deferred connections is that they may be connected to from threads other
than main.

In "Multiple Signals with Evaluator" test two threads tried to do a
deferred connection at the same time and apparently this caused race
condition in creation of the evaluator.

Now we unconditionally create the connection evaluator in the event
loop. This way, we know that the evaluator is fully created by the end
of the event loop construction. Our assumption is that user shouldn't
use the evaluator from another thread before the loop is fully
constructed.