Add clang-tidy on Windows on GitHub Actions
Switched back to turtlesec-no/get-ninja for getting ninja as
aseprite/get-ninja doesn't set the path to ninja in a way that makes
it usable by CMake on Windows.

It puts /c/Users/runneradmin/.local/bin into $PATH regardless if we're
running bash or pwsh. And for pwsh it's obviously not a correct path.

Even if I forced all the steps to use bash, CMake still couldn't find
Ninja for some reason.

Paths in CMake invocation don't have slashes now so they're more
platform-agnostic.