Don't fail fast
It's a nightly, better use the time to check the whole thing on all
configurations.

Unfortunately, with "continue-on-error: true" GitHub actions will still
mark the workflow run as successful even if all the steps fail. However,
we'd like to run Cppcheck even if clang-tidy gets errors.

Not possible, it's a limitation of GitHub actions documented in:
 - https://github.com/actions/runner/issues/2347
 - https://github.com/orgs/community/discussions/15452

GitHub folks collected feedback for this over a year ago:
https://github.com/orgs/community/discussions/15452#discussioncomment-5184613

but it's still not fixed.