Don't defer logger creation in platform classes
That posed a problems on how should we log in these plaform classes
constructors? We'd to log differently on construction. Plus, there would
be a contract requiring that all functions that log with SPDLOG_LOGGER_
has to happen after logger initialization happened or check (like
~LinuxXcbPlatformIntegration did).

However, it's not really needed. We _can_ initialize the loggers in
constructors. These loggers don't use the default logger that is
created by CoreApplication and the user-provided factory function
should also already be there because there's no way to set the
logger factory between CoreApplication constructor start and
AbstractPlatformIntegration::init() call that executes the delayed
logger init.