Rename MqttLib to MqttManager
MqttLib class is responsible for orchestrating
different aspects of MQTT:
- creating MqttClient instances
- init/deinit of mosquitto library
It will soon also own a logger instance for
the MQTT system and provide that to the
individual client instances.

Therefore we think MqttManager is a more
descriptive class name.
The fact that some of the functionality is
based on library functions is rather an
implementation detail worth hiding, than a
reason to keep the `Lib` suffix.