Platform application data and assets directories
To the core application class, add the ability to find Directories
necessary for storing applicatuion data files and assets. Calculate
the locations of these based on application metadata, as need be.

On Android, assets must be dealt with by different API from regular
files, so use a special indicator for the asset directory, which will
be dealt with in a following commit.

Change-Id: Idb63275e1991d1aa813f47f3a6431455f99666f6