Platform application data and assets directories
To the core application class, add the ability to find Directories
necessary for storing application data files and assets. Calculate
the locations of these based on application metadata, as need be.
Include the ability to switch between local/internal and roaming/
external appdata folders on Android and Windows.

On Android, assets must be dealt with by different API from regular
files, so add a common storage type element to files and directories.
Dealing with this on Android will be implemented in a subsequent
commit.

Change-Id: Idb63275e1991d1aa813f47f3a6431455f99666f6