Rework the websocket connection logic
To correctly process any initial websocket frames that may be included
in the same TCP packet as the websocket upgrade headers.

When the parser pauses for the upgrade, we store the offset in the
TCP data and then pass the remaining data to the websocket client.

We then use a deferred lambda to emit the connected signal and
then process any initial messages in that order.

Using the deferred signal emission means we can be sure that the
socket's signals have been disconnected from the HttpClient's slots.