Use proper hInstance when registering window class
We passed nullptr as hInstance and this could potentially be bad
if two instances of KDUtils were linked to the application
(e.g. plugins).

Unlikely but let's do it right. Current module is obtained through
a trick where a function passes itself to GetModuleHandleEx to make
sure we get the module handle of the current dll.

See: https://devblogs.microsoft.com/oldnewthing/20110715-00/?p=10133